fix automated migration
