+++ internal-enrichment/recordedfuture-enrichment/src/main.py
@@ -1,3 +1,14 @@
+"""
+############################## TERMS OF USE ####################################
+# The following code is provided for demonstration purposes only, and should   #
+# not be used without independent verification. Recorded Future makes no       #
+# representations or warranties, express, implied, statutory, or otherwise,    #
+# regarding this code, and provides it strictly "as-is".                       #
+# Recorded Future shall not be liable for, and you assume all risk of          #
+# using the foregoing.                                                         #
+################################################################################
+"""
+
 import traceback
 
 from pycti import OpenCTIConnectorHelper
