--- internal-enrichment/recordedfuture-enrichment/src/rflib/_version.py
@@ -1,12 +1,0 @@
-"""
-############################## TERMS OF USE ####################################
-# The following code is provided for demonstration purposes only, and should   #
-# not be used without independent verification. Recorded Future makes no       #
-# representations or warranties, express, implied, statutory, or otherwise,    #
-# regarding this code, and provides it strictly "as-is".                       #
-# Recorded Future shall not be liable for, and you assume all risk of          #
-# using the foregoing.                                                         #
-################################################################################
-"""
-
-__version__ = "2.0.0"
