+++ internal-enrichment/recordedfuture-enrichment/src/rflib/use_cases/enrich_vulnerability.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from connectors_sdk.models.octi import (
+from connectors_sdk.models import (
     AttackPattern,
     BaseIdentifiedEntity,
     ExternalReference,
