+++ internal-enrichment/recordedfuture-enrichment/tests/test_rflib/test_use_cases/test_enrich_vulnerability.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from unittest.mock import Mock
 
 import pytest
-from connectors_sdk.models.octi import Note, Software, Vulnerability
+from connectors_sdk.models import Note, Software, Vulnerability
 from rf_client.models import VulnerabilityEnrichment
 from rflib.use_cases.enrich_vulnerability import VulnerabilityEnricher
