+++ internal-enrichment/recordedfuture-enrichment/tests/tests_connector/test_settings.py
@@ -14,15 +14,19 @@ from rflib import ConnectorSettings
                 "connector": {
                     "id": "connector-id",
                     "name": "Test Connector",
-                    "scope": "test, connector",
+                    "scope": "ipv4-addr, ipv6-addr, domain-name, url, stixfile, vulnerability",
                     "log_level": "error",
                     "auto": True,
                 },
                 "recorded_future": {
                     "token": "SecretStr",
                     "create_indicator_threshold": 42,
-                    "info_max_tlp": "str",
-                    "vulnerability_enrichment_optional_fields": "str",
+                    "info_max_tlp": "TLP:CLEAR",
+                    "vulnerability_enrichment_optional_fields": [
+                        "aiInsights",
+                        "cpe",
+                        "risk",
+                    ],
                 },
             },
             id="full_valid_settings_dict",
@@ -33,8 +37,12 @@ from rflib import ConnectorSettings
                 "recorded_future": {
                     "token": "SecretStr",
                     "create_indicator_threshold": 42,
-                    "info_max_tlp": "str",
-                    "vulnerability_enrichment_optional_fields": "str",
+                    "info_max_tlp": "TLP:CLEAR",
+                    "vulnerability_enrichment_optional_fields": [
+                        "aiInsights",
+                        "cpe",
+                        "risk",
+                    ],
                 },
             },
             id="minimal_valid_settings_dict",
@@ -76,14 +84,18 @@ def test_settings_should_accept_valid_input(settings_dict):
                 "connector": {
                     "id": "connector-id",
                     "name": "Test Connector",
-                    "scope": "test, connector",
+                    "scope": "ipv4-addr, ipv6-addr, domain-name, url, stixfile, vulnerability",
                     "log_level": "error",
                 },
                 "recorded_future": {
                     "token": "SecretStr",
                     "create_indicator_threshold": 42,
-                    "info_max_tlp": "str",
-                    "vulnerability_enrichment_optional_fields": "str",
+                    "info_max_tlp": "TLP:CLEAR",
+                    "vulnerability_enrichment_optional_fields": [
+                        "aiInsights",
+                        "cpe",
+                        "risk",
+                    ],
                 },
             },
             "opencti.token",
@@ -95,14 +107,18 @@ def test_settings_should_accept_valid_input(settings_dict):
                 "connector": {
                     "id": 123456,
                     "name": "Test Connector",
-                    "scope": "test, connector",
+                    "scope": "ipv4-addr, ipv6-addr, domain-name, url, stixfile, vulnerability",
                     "log_level": "error",
                 },
                 "recorded_future": {
                     "token": "SecretStr",
                     "create_indicator_threshold": 42,
-                    "info_max_tlp": "str",
-                    "vulnerability_enrichment_optional_fields": "str",
+                    "info_max_tlp": "TLP:CLEAR",
+                    "vulnerability_enrichment_optional_fields": [
+                        "aiInsights",
+                        "cpe",
+                        "risk",
+                    ],
                 },
             },
             "connector.id",
